Corrected some warnings given by Visual Studio.
- Also, suppressed some warnings that make sense, but are too annoying.
- C4100 is annoying when there are callback functions that don't need required arguments. Removing their names and comments makes the callback connection harder to understand, and if one day the argument ends up being needed, it'll be trickier.
- C4458 is annoying when you have constructors that obviously want a value just to set one of their members, but both happen to share the same name because it's the most convenient name.

+++ Source/source/controls.cpp
@@ -680,7 +680,7 @@ void gameplay_state::handle_button(
             do {
                 next_maturity =
                     (size_t) sum_and_wrap(
-                        next_maturity,
+                        (int) next_maturity,
                         (button == BUTTON_NEXT_MATURITY ? 1 : -1),
                         N_MATURITIES
                     );
