+++ Source/source/game_states/gameplay/hud.cpp
@@ -923,9 +923,9 @@ hud_struct::hud_struct() :
         if(game.spray_types.size() >= 3) {
             prev_spray_idx =
                 sum_and_wrap(
-                    game.states.gameplay->selected_spray,
+                    (int) game.states.gameplay->selected_spray,
                     -1,
-                    game.spray_types.size()
+                    (int) game.spray_types.size()
                 );
         }
         if(prev_spray_idx == INVALID) return;
@@ -1777,7 +1777,7 @@ void hud_struct::tick(const float delta_t) {
             size_t l_idx =
                 (size_t) sum_and_wrap(
                     (int) game.states.gameplay->cur_leader_nr,
-                    l,
+                    (int) l,
                     (int) game.states.gameplay->available_leaders.size()
                 );
             l_ptr = game.states.gameplay->available_leaders[l_idx];
@@ -1902,9 +1902,9 @@ void hud_struct::tick(const float delta_t) {
     if(game.spray_types.size() >= 3) {
         next_spray_idx =
             sum_and_wrap(
-                game.states.gameplay->selected_spray,
+                (int) game.states.gameplay->selected_spray,
                 1,
-                game.spray_types.size()
+                (int) game.spray_types.size()
             );
     }
     spray_icon_mgr.update(
